newsportal/requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
django-allauth==0.58.2
django-apscheduler==0.6.2


